requirements.txt
streamlit
python-docx
pandas
numpy
scikit-learn
